# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Api",
        "description": "Test api",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@instafxgroup.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://test1210/api/v1",
            "description": "Dev API server"
        }
    ],
    "paths": {
...
        "/payment/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Payment info",
                "description": "Payment info",
                "operationId": "payment.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment unfo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/payment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update payment",
                "description": "Update payment",
                "operationId": "payment.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payment data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "description": "Amount",
                                        "type": "numeric",
                                        "format": "double",
                                        "example": "50.50",
                                        "nullable": "true"
                                    },
                                    "currency": {
                                        "description": "Cuttency",
                                        "type": "string",
                                        "example": "RUB",
                                        "nullable": "true"
                                    },
                                    "ticket": {
                                        "description": "Ticket",
                                        "type": "integer",
                                        "example": "123456",
                                        "nullable": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete payment",
                "description": "Delete payment",
                "operationId": "payment.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destroy result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There are some active events in this stream OR Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "error_response": {
                "title": "Error response",
                "properties": {
                    "error": {
                        "properties": {
                            "app": {
                                "type": "string",
                                "example": "Some Api"
                            },
                            "code": {
                                "type": "integer",
                                "example": "400"
                            },
                            "msg": {
                                "type": "string",
                                "example": "Bad request"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "user": {
                "title": "User data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "verified": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null",
                        "nullable": "true"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-04T16:33:11.000000Z"
                    },
                    "updated": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-04T16:33:11.000000Z"
                    }
                },
                "type": "object"
            },
            "payment": {
                "title": "Payment data",
                "properties": {
                    "payment_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "curency": {
                        "type": "string",
                        "example": "USD"
                    },
                    "amount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "50.50"
                    },
                    "amount_in_rub": {
                        "type": "numeric",
                        "format": "float",
                        "example": "5050"
                    },
                    "ticket": {
                        "type": "integer",
                        "example": "123456",
                        "nullable": "true"
                    },
                    "payer": {
                        "$ref": "#/components/schemas/user"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Your token must be provided by Admin",
                "name": "Api Client",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}